Fix bug where enclosing-braces would destroy commands (#127)
The enclosing-braces option wraps the entire value for specified fields in curly braces and removes any braces in the value (which are unnecessary because the entire value is now braced).

However, this was done with a regexp that removed braces even if they were part of a command. E.g. `The \textsf{Secret}` would become `{The \textsfSecret}`

This has been fixed by parsing the latex (using a very basic parser) and only removing braces that enclose a block, not a command.